Remove Object3DUtils references and related code
