+++ packages/spatial/src/transform/TransformModule.ts
@@ -1,5 +1,4 @@
 export { TransformComponent } from './components/TransformComponent'
 export { MeshBVHSystem } from './functions/MeshBVHSystem'
-export { Object3DUtils } from './Object3DUtils'
 export { SpawnPoseState } from './SpawnPoseState'
 export { TransformSystem } from './systems/TransformSystem'
